Убран вип чат. Сделана страница чата. Поправлены стили бургер-меню. Мобильные стили поправлены. Почищен код. Убрана модель вип-чата
