+++ backend/.env.prod.example
@@ -12,9 +12,6 @@ EMAIL_HOST_PASSWORD='uxftjmkthdxbgfsj'
 DEFAULT_FROM_EMAIL=koltsovaecoprint@yandex.ru
 ADMIN_EMAIL=kumaradji@me.com
 
-VIP_TELEGRAM_BOT_TOKEN=7359625002:AAEkb7K0ZuK1TnxRNwuR1lv_L-CWPTfaN6E
-VIP_TELEGRAM_CHAT_ID=707268574
-
 GENERAL_TELEGRAM_BOT_TOKEN=7289275372:AAGDhyO9xRWsO2s55qfcEjr-b0RKJXA2QMM
 GENERAL_TELEGRAM_CHAT_ID=707268574
