+++ backend/main/models.py
@@ -82,16 +82,6 @@ class Message(models.Model):
         return f"{self.user_name}: {self.content}"
 
 
-class VIPMessage(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    content = models.TextField()
-    timestamp = models.DateTimeField(auto_now_add=True)
-    is_admin = models.BooleanField(default=False)
-
-    def __str__(self):
-        return f"{self.user}: {self.content}"
-
-
 class Category(models.Model):
     name = models.CharField(max_length=255)
