+++ backend/main/signals.py
@@ -1,5 +1,4 @@
 # signals.py
-
 import os
 import shutil
 from django.db.models.signals import post_save, post_delete
