+++ frontend/package.json
@@ -22,6 +22,7 @@
     "react-select": "^5.8.0",
     "react-slick": "^0.30.2",
     "sass": "^1.72.0",
+    "shadcn-ui": "^0.9.2",
     "slick-carousel": "^1.8.1",
     "swiper": "^11.1.4",
     "web-vitals": "^2.1.4",
