--- frontend/src/components/ChatButton/ModalChat/ModalChat.module.scss
@@ -1,148 +1,0 @@
-@import '../../../styles/media';
-@import '../../../styles/variables';
-@import '../../../styles/mixins';
-
-.modalOverlay {
-	position: fixed;
-	top: 0;
-	left: 0;
-	width: 100%;
-	height: 100%;
-	background-color: rgba(0, 0, 0, 0.5);
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	z-index: 1000;
-}
-
-.modal {
-	background-color: var(--color-main);
-	padding: 20px;
-	border-radius: 8px;
-	position: relative;
-	width: 700px;
-	max-width: 90%;
-	display: flex;
-	flex-direction: column;
-}
-
-.modalHeader {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	margin-bottom: 15px;
-}
-
-.chatWindow {
-	flex: 1;
-	overflow-y: auto;
-	max-height: 500px;
-	border: 1px solid #ddd;
-	padding: 10px;
-	border-radius: 4px;
-	margin-bottom: 10px;
-}
-
-.message {
-	margin-bottom: 10px;
-	padding: 10px;
-	border-radius: 20px;
-	max-width: 60%;
-}
-
-.userMessage {
-	background-color: var(--color-light-green);
-	text-align: left;
-}
-
-.adminMessage {
-	background-color: var(--color-light-blue);
-	text-align: right;
-	margin-left: auto;
-}
-
-.timestamp {
-	font-size: 0.75em;
-	color: #777;
-	margin-top: 5px;
-	text-align: right;
-}
-
-.chatInput {
-	display: flex;
-	flex-direction: column;
-}
-
-.modal__button {
-	@include button-style
-}
-
-textarea {
-	width: 100%;
-	height: 50px;
-	padding: 10px;
-	border: 1px solid #ddd;
-	border-radius: 4px;
-	resize: none;
-	margin-bottom: 10px;
-}
-
-.modalButton {
-	width: 135px;
-	height: 30px;
-	font-family: 'Iosevka', sans-serif;
-	cursor: pointer;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	border: none;
-	border-radius: 4px;
-	font-size: 1rem;
-	color: var(--color-light-main);
-	background-color: var(--color-red);
-	transition: background-color 0.3s ease;
-
-	&:hover {
-		background-color: var(--color-blue);
-	}
-}
-
-.authRequired {
-	text-align: center;
-	font-size: 18px;
-}
-
-.modalButtonSend {
-	width: 135px;
-	height: 30px;
-	font-family: 'Iosevka', sans-serif;
-	cursor: pointer;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	border: none;
-	border-radius: 4px;
-	font-size: 1rem;
-	background-color: var(--color-blue);
-	color: var(--color-light-main);
-	transition: background-color 0.3s ease;
-
-	&:hover {
-		background-color: var(--color-green);
-	}
-}
-
-.authRequired p {
-	font-family: Iosevka, sans-serif;
-	margin-bottom: 20px;
-}
-
-.closeButton {
-	position: absolute;
-	top: 10px;
-	right: 10px;
-	background: none;
-	border: none;
-	font-size: 24px;
-	cursor: pointer;
-}
