+++ frontend/src/components/Header/Header.module.scss
@@ -39,12 +39,11 @@ header {
 	}
 
 	@include tablet {
-		padding: 0 12px;
-		height: 73px;
+		height: 60px;
 	}
 
 	@include mobile {
-		padding: 0 10px;
+		padding: 0 20px;
 	}
 }
 
@@ -59,7 +58,8 @@ header {
 	}
 
 	@include mobile {
-		width: 70px;
+		width: 80px;
+		z-index: 2001;
 	}
 }
 
@@ -84,11 +84,11 @@ header {
 	width: 765px;
 	max-width: 100%;
 	height: 100vh;
-	background-color: var(--color-red);
+	background-color: var(--color-light-main);
 	display: flex;
 	flex-direction: column;
 	align-items: center;
-	padding-top: 120px;
+	padding-top: 100px;
 	z-index: 2000;
 	overflow-y: auto;
 	pointer-events: auto;
@@ -100,13 +100,17 @@ header {
 	.footer__social {
 		display: flex;
 		justify-content: center;
-		gap: 20px;
-		margin-top: 20px;
+		gap: 30px;
+		margin-top: 100px;
 		margin-bottom: 30px;
 
+		@include mobile {
+			gap: 15px;
+		}
+
 		a img {
-			width: 60px;
-			height: 60px;
+			width: 50px;
+			height: 50px;
 		}
 
 		@include mobile-small {
@@ -140,10 +144,17 @@ header {
 	@include button-header;
 
 	@include tablet {
+		width: 70px;
+		height: 25px;
 		font-size: 13px;
-		width: 65px;
-		height: 24px;
 	}
+
+	@include mobile {
+		width: 65px;
+		height: 25px;
+		font-size: 12px;
+	}
+
 }
 
 .rightSection {
