+++ frontend/src/components/Header/Navbar/Navbar.module.scss
@@ -44,11 +44,11 @@
 	@include mobile {
 		margin-right: 0;
 		margin-bottom: 10px;
-		font-size: 14px;
+		font-size: 40px;
 	}
 
 	@include mobile-small {
-		font-size: 12px;
+		font-size: 20px;
 	}
 }
