+++ frontend/src/components/Header/UserBlock/UserBlock.module.scss
@@ -34,6 +34,22 @@
 	//min-width: 100px;
 	height: 32px;
 	align-items: center;
+
+	@include desktop {
+		height: 24px;
+	}
+
+	@include tablet {
+		justify-content: center;
+	}
+
+	@include mobile {
+		display: none;
+	}
+
+	@include mobile-small {
+		display: none;
+	}
 }
 
 .userInfo__details {
@@ -97,9 +113,6 @@
 	@include button-header;
 
 	@include mobile {
-		margin-left: auto;
-		width: 75px;
-		height: 30px;
-		font-size: 14px;
+		display: none;
 	}
 }
