+++ frontend/src/pages/MarenGarden/MarenGarden.jsx
@@ -2,18 +2,18 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { motion } from 'framer-motion';
 import styles from './MarenGarden.module.scss';
-import VIPChatCard from '../MarenGarden/VIPModalChat/VIPChatCard/VIPChatCard';
 
 const MarenGarden = ({ marenGardenChapters = [] }) => {
   return (
     <div className={styles.masterclass}>
       <h2>Мастер-класс "Цветной фон".</h2>
-      <p>Благодарю вас за то что вы обратили внимание на моё творчество и приобрели этот мастер-класс. Надеюсь это вдохновило вас на создание прекрасных и неповторимых шедевров, создаваемых вместе с природой.</p>
-      <p>Если у вас остались какие-то вопросы, пожалуйста напишите мне. Я обязательно на них отвечу.</p>
+      <p>Благодарю вас за то, что вы обратили внимание на моё творчество и приобрели этот мастер-класс. Надеюсь, это вдохновило вас на создание прекрасных и неповторимых шедевров, создаваемых вместе с природой.</p>
+      <p>Если у вас остались какие-то вопросы, пожалуйста, напишите мне. Я обязательно на них отвечу.</p>
       <p>Этот мастер-класс лишь "вершина айсберга" бездонного океана экопринта.</p>
       <p>Экспериментируйте, творите, а я всегда помогу вам в этом.</p>
+
+      {/* Секция с главами мастер-класса */}
       <div className={styles.chapters}>
-        <VIPChatCard />
         {marenGardenChapters.map((chapter, index) => (
           <motion.div
             key={chapter.id}
@@ -32,6 +32,19 @@ const MarenGarden = ({ marenGardenChapters = [] }) => {
             </div>
           </motion.div>
         ))}
+      </div>
+
+      {/* Вкладка Часто задаваемые вопросы */}
+      <div className={styles.faqSection}>
+        <h3>Часто задаваемые вопросы</h3>
+        <ul>
+          <li><strong>Вопрос 1:</strong> Как долго доступен мастер-класс?</li>
+          <li><strong>Ответ:</strong> Доступ к мастер-классу предоставляется бессрочно.</li>
+          <li><strong>Вопрос 2:</strong> Могу ли я скачать видео мастер-класса?</li>
+          <li><strong>Ответ:</strong> Нет, видео доступны только для просмотра онлайн.</li>
+          <li><strong>Вопрос 3:</strong> Как задать вопрос автору?</li>
+          <li><strong>Ответ:</strong> Вы можете связаться со мной через форму обратной связи на сайте или через наш закрытый Telegram-канал.</li>
+        </ul>
       </div>
     </div>
   );
