+++ frontend/src/pages/MarenGarden/MarenGarden.module.scss
@@ -224,3 +224,45 @@
 		}
 	}
 }
+
+.telegramButtonContainer {
+	text-align: center;
+	margin-top: 30px;
+	padding: 20px;
+	background-color: #f1f1f1;
+}
+
+.telegramButton {
+	display: inline-block;
+	background-color: #0088cc;
+	color: white;
+	padding: 10px 20px;
+	text-decoration: none;
+	border-radius: 5px;
+	font-weight: bold;
+	transition: background-color 0.3s;
+}
+
+.telegramButton:hover {
+	background-color: #0077aa;
+}
+
+.faqSection {
+	margin-top: 40px;
+	padding: 20px;
+	background-color: #f9f9f9;
+}
+
+.faqSection h3 {
+	margin-bottom: 20px;
+}
+
+.faqSection ul {
+	list-style-type: none;
+	padding: 0;
+}
+
+.faqSection li {
+	margin-bottom: 10px;
+}
+
