+++ frontend/src/pages/ShopPage/ShopPage.module.scss
@@ -203,32 +203,30 @@
 	}
 
 	.burgerButton {
-		background: var(--color-blue);
-		//opacity: 0.9;
+		background: var(--color-light-main);
 		cursor: pointer;
-		font-size: 1.0rem;
-		color: var(--color-main);
+		font-size: 1.1rem;
+		color: var(--color-black);
 		border-radius: 5px;
-		padding: 10px;
-		margin: 70px 0;
+		padding: 8px;
+		margin: 50px 15px;
 		text-align: center;
 		box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
 	}
 
 	.burgerMenuOpen {
-		background-color: var(--color-blue);
+		background-color: var(--color-light-main);
 		opacity: 0.9;
-		padding: 10px;
+		padding: 8px;
 		border-radius: 5px;
 		box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-		margin-top: 10px;
 
 		ul {
 			list-style: none;
 			padding: 0;
 			margin: 0;
-			font-size: 1.0rem;
-			color: var(--color-main);
+			font-size: 1.1rem;
+			color: var(--color-black);
 
 			li {
 				margin-bottom: 10px;
